Move  `/src/demos/*` to `/demos/*`
