+++ demos/multiroot-react/App.tsx
