+++ demos/multiroot-react/MultiRootEditor.tsx
