+++ demos/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "types": [
+      "../vite-env.d.ts"
+    ]
+  },
+  "include": [
+    "**/*"
+  ]
+}
