+++ vite-env.d.ts
@@ -1,0 +1,1 @@
+declare const __REACT_VERSION__: number;
